create json utils class
