+++ src/main/java/info/movito/themoviedbapi/TmdbChanges.java
@@ -1,10 +1,12 @@
 package info.movito.themoviedbapi;
 
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.time.temporal.ChronoUnit;
+
 import info.movito.themoviedbapi.model.changes.ChangesResultsPage;
 import info.movito.themoviedbapi.tools.ApiUrl;
 import info.movito.themoviedbapi.tools.TmdbException;
-
-import static info.movito.themoviedbapi.util.Utils.calculateDaysDifference;
 
 /**
  * The movie database api for changes. See the
@@ -93,5 +95,20 @@ public class TmdbChanges extends AbstractTmdbApi {
         apiUrl.addPage(page);
 
         return mapJsonResult(apiUrl, ChangesResultsPage.class);
+    }
+
+    /**
+     * Calculate the difference in days between two date strings.
+     *
+     * @param startDateString the start date string, in format: YYYY-MM-DD.
+     * @param endDateString   the end date string, in format: YYYY-MM-DD.
+     * @return the difference in days.
+     */
+    private static long calculateDaysDifference(String startDateString, String endDateString) {
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+        LocalDate startDate = LocalDate.parse(startDateString, formatter);
+        LocalDate endDate = LocalDate.parse(endDateString, formatter);
+
+        return ChronoUnit.DAYS.between(startDate, endDate);
     }
 }
