+++ src/main/java/info/movito/themoviedbapi/TmdbMovies.java
@@ -22,7 +22,7 @@ import info.movito.themoviedbapi.tools.ApiUrl;
 import info.movito.themoviedbapi.tools.RequestType;
 import info.movito.themoviedbapi.tools.TmdbException;
 import info.movito.themoviedbapi.tools.appendtoresponse.MovieAppendToResponse;
-import info.movito.themoviedbapi.util.Utils;
+import info.movito.themoviedbapi.util.JsonUtil;
 
 /**
  * The movie database api for movies. See the
@@ -325,7 +325,7 @@ public class TmdbMovies extends AbstractTmdbApi {
         HashMap<String, Object> body = new HashMap<>();
         body.put("value", rating);
 
-        String jsonBody = Utils.convertToJson(getObjectMapper(), body);
+        String jsonBody = JsonUtil.convertToJson(getObjectMapper(), body);
         return mapJsonResult(apiUrl, jsonBody, RequestType.POST, ResponseStatus.class);
     }
