+++ src/main/java/info/movito/themoviedbapi/util/JsonUtil.java
@@ -1,0 +1,26 @@
+package info.movito.themoviedbapi.util;
+
+import java.util.Map;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+/**
+ * Utility class.
+ */
+public final class JsonUtil {
+    private JsonUtil() {
+    }
+
+    /**
+     * Use Jackson to convert Map to json string.
+     */
+    public static String convertToJson(ObjectMapper objectMapper, Map<String, ?> map) {
+        try {
+            return objectMapper.writeValueAsString(map);
+        }
+        catch (JsonProcessingException jpe) {
+            throw new RuntimeException("json conversion failed", jpe);
+        }
+    }
+}
