+++ src/test/java/info/movito/themoviedbapi/TmdbMoviesTest.java
@@ -26,8 +26,8 @@ import info.movito.themoviedbapi.tools.TmdbException;
 import info.movito.themoviedbapi.tools.TmdbResponseCode;
 import info.movito.themoviedbapi.tools.appendtoresponse.MovieAppendToResponse;
 import info.movito.themoviedbapi.util.AbstractJsonMappingValidator;
+import info.movito.themoviedbapi.util.JsonUtil;
 import info.movito.themoviedbapi.util.TestUtils;
-import info.movito.themoviedbapi.util.Utils;
 import org.junit.jupiter.api.Test;
 
 import static info.movito.themoviedbapi.TmdbMovies.TMDB_METHOD_MOVIE;
@@ -355,7 +355,7 @@ public class TmdbMoviesTest extends AbstractTmdbApiTest<TmdbMovies> {
     public void testAddRating() throws IOException, TmdbException {
         HashMap<String, Object> requestBody = new HashMap<>();
         requestBody.put("value", 2.1);
-        String jsonBody = Utils.convertToJson(AbstractTmdbApi.getObjectMapper(), requestBody);
+        String jsonBody = JsonUtil.convertToJson(AbstractTmdbApi.getObjectMapper(), requestBody);
 
         String url = TMDB_API_BASE_URL + TMDB_METHOD_MOVIE + "/123/rating";
         String body = TestUtils.readTestFile("api_responses/movies/add_rating.json");
